pyproject.toml
[project]
name = "dagster_portada_project"
version = "0.0.8"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "dagster",
    "dagster_pyspark",
    "py_portada_data_layer @ git+https://github.com/portada-git/py_portada_data_layer",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver",
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

#[tool.dagster]
#module_name = "dagster_portada_project/definitions.py"
#working_directory = "."
#
#[tool.setuptools.packages.find]
#exclude=["dagster_portada_project_tests"]
